moved PNCounter replica methods fro cluster service to PNCounter
