+++ internal/cluster/cluster_service.go
@@ -19,10 +19,8 @@ package cluster
 import (
 	"context"
 	"fmt"
-	"math/rand"
 	"strings"
 	"sync"
-	"time"
 
 	pubcluster "github.com/hazelcast/hazelcast-go-client/cluster"
 	"github.com/hazelcast/hazelcast-go-client/internal/event"
@@ -32,8 +30,6 @@ import (
 	"github.com/hazelcast/hazelcast-go-client/internal/proto/codec"
 	"github.com/hazelcast/hazelcast-go-client/types"
 )
-
-var commonRand = rand.New(rand.NewSource(time.Now().UnixNano()))
 
 type Service struct {
 	logger            ilogger.Logger
@@ -99,15 +95,8 @@ func (s *Service) MemberAddrs() []pubcluster.Address {
 	return s.membersMap.MemberAddrs()
 }
 
-func (s *Service) RandomReplica(n int) (pubcluster.MemberInfo, bool) {
-	return s.membersMap.RandomReplica(n, nil)
-}
-
-func (s *Service) RandomReplicaExcluding(n int, excluded map[types.UUID]struct{}) (pubcluster.MemberInfo, bool) {
-	return s.membersMap.RandomReplica(n, func(mem *pubcluster.MemberInfo) bool {
-		_, found := excluded[mem.UUID]
-		return !found
-	})
+func (s *Service) OrderedMembers() []pubcluster.MemberInfo {
+	return s.membersMap.OrderedMembers()
 }
 
 func (s *Service) RefreshedSeedAddrs(clusterCtx *CandidateCluster) ([]pubcluster.Address, error) {
@@ -255,13 +244,12 @@ func (m *membersMap) MemberAddrs() []pubcluster.Address {
 	return addrs
 }
 
-// RandomReplica returns one of the replicas (first n data members).
-// filter, if provided should return true for considering the data member in the list of replicas.
-// Returns false if no suitable data member was found.
-func (m *membersMap) RandomReplica(n int, filter func(mem *pubcluster.MemberInfo) bool) (pubcluster.MemberInfo, bool) {
-	m.membersMu.RLock()
-	defer m.membersMu.RUnlock()
-	return randomReplica(m.orderedMembers, n, filter)
+func (m *membersMap) OrderedMembers() []pubcluster.MemberInfo {
+	m.membersMu.Lock()
+	members := make([]pubcluster.MemberInfo, len(m.orderedMembers))
+	copy(members, m.orderedMembers)
+	m.membersMu.Unlock()
+	return members
 }
 
 // addMember adds the given memberinfo if it doesn't already exist and returns true in that case.
@@ -302,29 +290,6 @@ func (m *membersMap) reset() {
 	m.addrToMemberUUID = map[pubcluster.Address]types.UUID{}
 	m.version = -1
 	m.membersMu.Unlock()
-}
-
-func randomReplica(members []pubcluster.MemberInfo, n int, filter func(mem *pubcluster.MemberInfo) bool) (pubcluster.MemberInfo, bool) {
-	if n > len(members) {
-		n = len(members)
-	}
-	if n == 0 {
-		return pubcluster.MemberInfo{}, false
-	}
-	// scans first n members, starting from idx, wrapping at n
-	idx := commonRand.Intn(n)
-	for i, mem := range members {
-		if mem.LiteMember {
-			continue
-		}
-		if i < idx {
-			continue
-		}
-		if filter == nil || filter(&mem) {
-			return mem, true
-		}
-	}
-	return pubcluster.MemberInfo{}, false
 }
 
 func (m *membersMap) logMembers(version int32) {
