+++ proxy_pn_counter.go
@@ -19,7 +19,9 @@ package hazelcast
 import (
 	"context"
 	"fmt"
+	"math/rand"
 	"sync"
+	"time"
 
 	"github.com/hazelcast/hazelcast-go-client/cluster"
 	"github.com/hazelcast/hazelcast-go-client/hzerrors"
@@ -30,6 +32,8 @@ import (
 	iproxy "github.com/hazelcast/hazelcast-go-client/internal/proxy"
 	"github.com/hazelcast/hazelcast-go-client/types"
 )
+
+var commonRand = rand.New(rand.NewSource(time.Now().UnixNano()))
 
 /*
 PNCounter is a PN (Positive-Negative) CRDT counter.
@@ -142,9 +146,9 @@ func (pn *PNCounter) crdtOperationTarget(excluded map[types.UUID]struct{}) (*clu
 		var target cluster.MemberInfo
 		var ok bool
 		if excluded == nil {
-			target, ok = pn.clusterService.RandomReplica(int(pn.maxReplicaCount))
+			target, ok = pn.randomReplica(int(pn.maxReplicaCount))
 		} else {
-			target, ok = pn.clusterService.RandomReplicaExcluding(int(pn.maxReplicaCount), excluded)
+			target, ok = pn.randomReplicaExcluding(int(pn.maxReplicaCount), excluded)
 		}
 		if !ok {
 			return nil, nil
@@ -216,10 +220,51 @@ func (pn *PNCounter) invokeOnMember(ctx context.Context, makeReq func(target typ
 	})
 }
 
+// randomReplica returns one of the replicas (first n data members).
+// Returns false if no suitable data member was found.
+func (pn *PNCounter) randomReplica(n int) (cluster.MemberInfo, bool) {
+	members := pn.clusterService.OrderedMembers()
+	return randomReplica(members, n, nil)
+}
+
+// randomReplicaExcluding returns one of the replicas (first n data members).
+// Members with UUIDs in the excluded set are not considered in the result.
+// Returns false if no suitable data member was found.
+func (pn *PNCounter) randomReplicaExcluding(n int, excluded map[types.UUID]struct{}) (cluster.MemberInfo, bool) {
+	members := pn.clusterService.OrderedMembers()
+	return randomReplica(members, n, func(mem *cluster.MemberInfo) bool {
+		_, found := excluded[mem.UUID]
+		return !found
+	})
+}
+
 func targetExcluded(target *cluster.MemberInfo, excludes map[types.UUID]struct{}) bool {
 	if excludes == nil {
 		return false
 	}
 	_, found := excludes[target.UUID]
 	return found
+}
+
+func randomReplica(members []cluster.MemberInfo, n int, filter func(mem *cluster.MemberInfo) bool) (cluster.MemberInfo, bool) {
+	if n > len(members) {
+		n = len(members)
+	}
+	if n == 0 {
+		return cluster.MemberInfo{}, false
+	}
+	// scans first n members, starting from idx, wrapping at n
+	idx := commonRand.Intn(n)
+	for i, mem := range members {
+		if mem.LiteMember {
+			continue
+		}
+		if i < idx {
+			continue
+		}
+		if filter == nil || filter(&mem) {
+			return mem, true
+		}
+	}
+	return cluster.MemberInfo{}, false
 }
